:muscle: Move `worktree.ts` into `git` module
