+++ denops/gin/core/util/main.ts
@@ -2,7 +2,7 @@ import type { Denops } from "https://deno.land/x/denops_std@v4.0.0/mod.ts";
 import * as unknownutil from "https://deno.land/x/unknownutil@v2.1.0/mod.ts";
 import * as buffer from "https://deno.land/x/denops_std@v4.0.0/buffer/mod.ts";
 import { expand } from "../../util/cmd.ts";
-import { findWorktreeFromDenops } from "../../util/worktree.ts";
+import { findWorktreeFromDenops } from "../../git/worktree.ts";
 
 export function main(denops: Denops): void {
   denops.dispatcher = {
