+++ denops/gin/feat/branch/command.ts
@@ -14,7 +14,7 @@ import {
   validateOpts,
 } from "https://deno.land/x/denops_std@v4.0.0/argument/mod.ts";
 import { normCmdArgs } from "../../util/cmd.ts";
-import { findWorktreeFromDenops } from "../../util/worktree.ts";
+import { findWorktreeFromDenops } from "../../git/worktree.ts";
 
 const allowedFlags = [
   "a",
