+++ denops/gin/feat/chaperon/command.ts
@@ -14,7 +14,7 @@ import {
 } from "https://deno.land/x/denops_std@v4.0.0/argument/mod.ts";
 import * as buffer from "https://deno.land/x/denops_std@v4.0.0/buffer/mod.ts";
 import { normCmdArgs } from "../../util/cmd.ts";
-import { findWorktreeFromDenops } from "../../util/worktree.ts";
+import { findWorktreeFromDenops } from "../../git/worktree.ts";
 import { exec as execEdit } from "../edit/command.ts";
 import { AliasHead, getInProgressAliasHead, stripConflicts } from "./util.ts";
