+++ denops/gin/feat/status/action.ts
@@ -5,7 +5,7 @@ import * as batch from "https://deno.land/x/denops_std@v4.0.0/batch/mod.ts";
 import * as buffer from "https://deno.land/x/denops_std@v4.0.0/buffer/mod.ts";
 import * as fn from "https://deno.land/x/denops_std@v4.0.0/function/mod.ts";
 import * as helper from "https://deno.land/x/denops_std@v4.0.0/helper/mod.ts";
-import { findWorktreeFromDenops } from "../../util/worktree.ts";
+import { findWorktreeFromDenops } from "../../git/worktree.ts";
 import { define, Range } from "../../core/action/action.ts";
 import { command as commandEdit } from "../edit/command.ts";
 import { command as commandDiff } from "../diff/command.ts";
