+++ denops/gin/git/worktree.ts
@@ -6,7 +6,7 @@ import * as path from "https://deno.land/std@0.171.0/path/mod.ts";
 import * as unknownutil from "https://deno.land/x/unknownutil@v2.1.0/mod.ts";
 import { GIN_BUFFER_PROTOCOLS } from "../global.ts";
 import { expand } from "../util/cmd.ts";
-import { find } from "../git/finder.ts";
+import { find } from "./finder.ts";
 
 /**
  * Find a git worktree from a suspected directory
